Directly deleting the GuiLayer or the World could cause it to crash due to the deletion being on a separate thread while the main thread could be everywhere in the loop You should definitely hand the deletion of object to the Engine's main thread

+++ include/Classes/Action.h
@@ -25,7 +25,7 @@ class Action
         }
 
         private:
-        int press_frame = -1;
+        long long press_frame = -1;
 
         std::unordered_set<SDL_KeyCode> m_keycodes;
         std::unordered_set<SDL_Scancode> m_scancodes;
